requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
easydict==1.9
graphviz==0.20
idna==3.3
imageio==2.19.2
joblib==1.1.0
networkx==2.8.2
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
tifffile==2022.5.4
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchinfo==1.6.6
torchvision==0.12.0+cpu
torchviz==0.0.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
